\name{.DollarNames-methods}
\docType{methods}
\alias{.DollarNames-methods}
\alias{.DollarNames,ANY,ANY-method}
\alias{.DollarNames,C++Object,character-method}
\title{Completion for C++ objects}
\description{
Completion for C++ objects
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY", pattern = "ANY")}}{
Standard method (from utils)
}

\item{\code{signature(x = "C++Object", pattern = "character")}}{
	Method for \linkS4class{"C++Object"} objects.
}
}}
\keyword{methods}

