\name{C++ObjectS3-class}
\Rdversion{1.1}
\docType{class}
\alias{C++ObjectS3-class}

\title{Class "C++ObjectS3"}
\description{deprecated class}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "C++ObjectS3" in the signature.
}
\examples{
showClass("C++ObjectS3")
}
\keyword{classes}
