\name{referenceMethods-methods}
\docType{methods}
\alias{referenceMethods,C++Class-method}
\title{generate reference methods proxies for C++ classes}
\description{
 generate reference methods proxies for C++ classes exposed
 by Rcpp modules
}
\section{Methods}{
\describe{

\item{\code{signature(classDef = "C++Class")}}{
generate reference methods proxies for C++ classes
}
}}
\keyword{methods}

