\name{.DollarNames-methods}
\docType{methods}
\alias{.DollarNames-methods}
\alias{.DollarNames,ANY-method}
\alias{.DollarNames,C++Object-method}
\alias{.DollarNames,Module-method}
\title{ completion }
\description{
 completion
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ANY")}}{}

\item{\code{signature(x = "C++Object")}}{
completes fields and methods of C++ objects
}

\item{\code{signature(x = "Module")}}{
completes functions and classes of modules
}
}}
\keyword{methods}

