\name{formals<--methods}
\docType{methods}
\alias{formals<--methods}
\alias{formals<-,C++Function-method}
\title{Set the formal arguments of a C++ function}
\description{
Set the formal arguments of a C++ function
}
\section{Methods}{
\describe{

\item{\code{signature(fun = "C++Function")}}{
Set the formal arguments of a C++ function
}
}}
\keyword{methods}
