\name{populate}
\alias{populate}
\title{
Populates a namespace or an environment with the content of a module
}
\description{
Populates a namespace or an environment with the content of a module
}
\usage{
populate(module, env)
}
\arguments{
  \item{module}{Rcpp module}
  \item{env}{environment or namespace}
}
\keyword{manip}

