\name{Rcpp-deprecated}
\alias{Rcpp-deprecated}
\title{Deprecated Functions in the Rcpp Package}
\description{
  These functions are provided for compatibility with older versions of the
  \pkg{Rcpp} package only, and may be removed in future versions.
}
\details{
  \code{loadRcppModules} calls should now be replaced by
  \code{\link{loadModule}} calls, one per Module.
}
