\name{RcppDemo}
\alias{RcppDemo}

\title{Run Rcpp demos Interactively}
\description{
  Permits the user to run all demos in Rcpp by selecting from a list.
}
\usage{RcppDemo()}
\details{
  Displays a list of all of the demos available for the package \code{Rcpp}
  and permits the user to select demos to run from the list.

  Entering 0 lists all of the demos along with a short description of
  what it does.
}
\author{Dominick Samperi}
\examples{
RcppDemo()
}
\keyword{misc}
