\name{RcppTemplateDemo}
\alias{RcppTemplateDemo}

\title{Run RcppTemplate demos Interactively}
\description{
  Permits the user to run all demos in RcppTemplate by selecting from a list.
}
\usage{RcppTemplateDemo()}
\details{
  Displays a list of all of the demos available for
  the package \code{RcppTemplate}
  and permits the user to select demos to run from the list.

  Entering 0 lists all of the demos along with a short description of
  what it does.
}
\author{Dominick Samperi}
\examples{

RcppTemplateDemo()

}
\keyword{misc}
