\name{RcppTemplateVersion}
\alias{RcppTemplateVersion}
\title{Display version info for package and for Rcpp/RcppTemplate}
\description{
  \code{RcppTemplateVersion} displays version information about the
  package, along with the version of Rcpp/RcppTemplate that was
  used to build the package.
}
\usage{
RcppTemplateVersion()
}
\author{Dominick Samperi}
\examples{
RcppTemplateVersion()
}

\keyword{models}
