\name{RcppVersion}
\alias{RcppVersion}
\title{Display version info for package and for Rcpp}
\description{
  \code{RcppVersion} displays version information about the
  package, along with the version of Rcpp that was 
  used to build the package.
}
\usage{
RcppVersion()
}
\author{Dominick Samperi}
\examples{
RcppVersion()
}
\keyword{models}
