\name{C++ObjectS3-class}
\Rdversion{1.1}
\docType{class}
\alias{C++ObjectS3-class}

\title{Class "C++ObjectS3"}
\description{
	This class is used to allow S3 dispatch of \code{\link[utils]{.DollarNames}}
	which is an S3 generic. Rcpp defines an S3 method for the 
	\code{"C++ObjectS3"} which dispatches internally to the appropriate method. 
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
}
\section{Methods}{
No methods defined with class "C++ObjectS3" in the signature.
}
\references{
R-devel thread and workaround from John Chambers: 
\url{http://permalink.gmane.org/gmane.comp.lang.r.devel/24524}
}
\keyword{classes}
