\name{C++OverloadedMethods-class}
\Rdversion{1.1}
\docType{class}
\alias{C++OverloadedMethods-class}

\title{Class "C++OverloadedMethods"}
\description{
Set of C++ methods
}
\section{Extends}{
Class \code{"\linkS4class{envRefClass}"}, directly.
Class \code{"\linkS4class{.environment}"}, by class "envRefClass", distance 2.
Class \code{"\linkS4class{refClass}"}, by class "envRefClass", distance 2.
Class \code{"\linkS4class{environment}"}, by class "envRefClass", distance 3, with explicit coerce.
Class \code{"\linkS4class{refObject}"}, by class "envRefClass", distance 3.
}
\keyword{classes}
\section{Fields}{
  \describe{
    \item{\code{pointer}:}{Object of class \code{externalptr} pointer to the internal structure that represents the set of methods }
    \item{\code{class_pointer}:}{Object of class \code{externalptr} pointer to the internal structure that models the related class }
  }
}
