\name{RcppVersion}
\alias{RcppVersion}
\title{Display version info for package and for Rcpp API}
\description{
  \code{RcppVersion} displays version information about the
  package, along with the version of Rcpp that was 
  used to build the package.
}
\usage{
RcppVersion()
}
\author{Dominick Samperi wrote most of Rcpp during 2005 and 2006.  Dirk
  Eddelbuettel made some additions, and became maintainer in 2008.}
\examples{
RcppVersion()
}
\keyword{interface}
